refactor(workflow-assistant): address self-review findings
- Move the applyWorkflowAssistantToolCalls reducer (business logic) out of the
  contract/ layer — which is schemas/types only — into platform-core beside the
  other workflow-graph helpers, along with the WorkflowAssistantToolCall union
  it needs. Repoint the UI, handler, and test to platform-core; contract keeps
  re-exporting the type for a single import site.
- Reconcile PLAN-0011 / ADR-0011 with what shipped: the prompt is sourced from
  the three embedded authoring docs (not getWorkflowAuthorableJsonSchema/
  loadWorkflowExamples), the chat shell is a fresh session-free build modelled
  on ChatCoworkView, and list_models + per-call audit logging are noted.
- Log the save->Runs redirect under CHANGELOG Changed.
- Widen the tool-definition entries cast from typeof AddStepToolSchema to
  z.ZodType (Update/Remove were mistyped as the Add schema).
