+++ docs/adr/PLAN-0012.md
@@ -5,11 +5,11 @@ Companion to [`0012-unified-canvas-first-workflow-designer.md`](./0012-unified-c
 ## Steps
 
 - **Tool schema**: Zod schema for canvas-mutation tool calls — `add_step`, `update_step`, `remove_step` (snake_case tool names). Transitions are expressed through `add_step`'s `insertAfterId`/`insertBeforeId` (place a step in the flow in one call) rather than a separate `add_transition` tool.
-- **System prompt**: adapted from `workflow-designer.wd.json`'s `design` step prompt + `getWorkflowAuthorableJsonSchema()` + a few `docs/workflow-examples/` entries for few-shot grounding.
-- **Headless handler + contract**: one handler (`assistant.ask`) — takes `{messages, model, workflowDefinition}` and runs a bounded tool-call loop: each round calls `callOpenRouter()` with the prompt + tool schema, applies any tool calls, and server-side-validates the resulting graph/references, re-prompting the model when it leaves the graph incomplete. Returns a text reply and/or the applied tool calls. No engine, no session persistence yet.
+- **System prompt**: assistant-specific behaviour (routing, tool mechanics, conversational style) plus the three authoring docs — `docs/how-to-create-workflow.md`, `docs/workflow-capabilities.md`, `docs/workflow-authoring-golden-rules.md` — embedded verbatim as the capability & authoring reference. Baked in at build time via `scripts/gen_assistant_docs.py` → a committed generated module (`_lib/embedded-workflow-docs.generated.ts`), since platform-api is consumed as source across runtimes with no single bundler; a drift test fails CI if the docs change without regenerating.
+- **Headless handler + contract**: one handler (`assistant.ask`) — takes `{messages, model, workflowDefinition}` and runs a bounded tool-call loop: each round calls `callOpenRouter()` with the prompt + tool schema, applies any tool calls, and server-side-validates the resulting graph/references, re-prompting the model when it leaves the graph incomplete. Returns a text reply and/or the applied tool calls, and writes a `workflow_assistant.prompt` audit entry per call. No engine, no conversation/session state persisted yet.
 - **Route + CLI**: Next.js route adapter wiring the handler, plus `mediforce.assistant.ask()` client method and CLI command (AGENTS.md §4, dogfood the CLI).
-- **Model config**: drop `ModelPicker` into the AI pane header, default to the existing `DEFAULT_MODEL` convention (`anthropic/claude-sonnet-4`).
-- **Chat UI**: swap the "To be implemented" placeholder for the reused `ChatCoworkView` pieces (message list, bubble, input) — local component state only, no artifact panel, no polling.
+- **Model config**: drop `ModelPicker` into the AI pane header, default to the existing `DEFAULT_MODEL` convention (`anthropic/claude-sonnet-4`). Also expose a `list_models` tool so the assistant can resolve a real model ID when the user describes one in words ("cheap", "your best model").
+- **Chat UI**: swap the "To be implemented" placeholder for a fresh, session-free chat shell (message list, bubble, input) modelled on `ChatCoworkView` — local component state only, no artifact panel, no polling. (`ChatCoworkView` is session/instance-bound, so it's the model, not a direct import.)
 - **Wire tool calls → canvas**: when the assistant returns tool calls, apply them to the canvas state. Shipped as a shared pure reducer (`applyWorkflowAssistantToolCalls`) used by both the UI and the handler's server-side validation. Collapsing this with the interactive `addStep`/`removeStep` path so both go through one primitive is tracked for the tools-paradigm overhaul.
 - **L3 test**: one API-level test for the handler — given a prompt (mocked LLM response), assert it either produces the right step mutation or a plain-text reply.
 - **Manual verification**: describe a workflow → it builds steps; ask a question → plain answer; ambiguous request → clarifying question back.
