+++ packages/platform-api/src/contract/index.ts
@@ -132,18 +132,11 @@ export {
 export {
   WorkflowAssistantMessageSchema,
   AskWorkflowAssistantInputSchema,
-  WorkflowAssistantToolCallSchema,
   AskWorkflowAssistantOutputSchema,
   type AskWorkflowAssistantInput,
   type WorkflowAssistantToolCall,
   type AskWorkflowAssistantOutput,
 } from './workflow-assistant';
-
-export {
-  applyWorkflowAssistantToolCalls,
-  type ToolCallOutcome,
-  type ApplyToolCallsResult,
-} from './workflow-assistant-apply';
 
 export {
   DockerImageInfoSchema,
