+++ packages/platform-core/src/index.ts
@@ -265,11 +265,16 @@ export {
   ListModelsToolSchema,
   WORKFLOW_ASSISTANT_TOOLS,
   WORKFLOW_ASSISTANT_DEFAULT_MODEL,
+  WorkflowAssistantToolCallSchema,
+  applyWorkflowAssistantToolCalls,
   type AddStepTool,
   type UpdateStepTool,
   type RemoveStepTool,
   type ListModelsTool,
   type WorkflowAssistantToolName,
+  type WorkflowAssistantToolCall,
+  type ToolCallOutcome,
+  type ApplyToolCallsResult,
 } from './schemas/index';
 
 // Interfaces (repository and service contracts)
